requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via flask
contourpy==1.0.7
    # via matplotlib
cycler==0.11.0
    # via matplotlib
easy-trilateration==0.1.4
    # via -r requirements.in
flask==2.2.3
    # via -r requirements.in
fonttools==4.39.3
    # via matplotlib
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
kiwisolver==1.4.4
    # via matplotlib
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.7.1
    # via
    #   -r requirements.in
    #   easy-trilateration
numpy==1.24.2
    # via
    #   contourpy
    #   matplotlib
    #   scipy
packaging==23.0
    # via matplotlib
pillow==9.4.0
    # via
    #   -r requirements.in
    #   matplotlib
ping3==4.0.4
    # via -r requirements.in
pygame==2.3.0
    # via -r requirements.in
pyparsing==3.0.9
    # via matplotlib
python-dateutil==2.8.2
    # via matplotlib
requests==2.28.2
    # via -r requirements.in
scipy==1.10.1
    # via easy-trilateration
six==1.16.0
    # via python-dateutil
urllib3==1.26.15
    # via requests
werkzeug==2.2.3
    # via flask
